feat: switch function bundling to esbuild
